Add Table components for farm input product sales report (#342)
* feat: Add Table components for farm input product sales report

The code changes include adding three new files: TableBody.tsx, TableHead.tsx, and TableFooter.tsx. These files contain the components for the table body, table head, and table footer of the farm input product sales report table. The components are responsible for rendering the data, headers, and footer of the table respectively. This addition improves the organization and structure of the code for the report table.

* feat: Update npm dependencies to latest versions

This commit updates several npm dependencies to their latest stable versions. The updated dependencies include:
- @sentry/nextjs: from version 8.26.0 to 8.27.0
- @serwist/next: from version 9.0.5 to 9.0.6
- moment: removed from dependencies

This update ensures that the project is using the most recent versions of these dependencies, which may include bug fixes, performance improvements, and new features.