Add pagination functionality to ProductPicker component (#343)
* feat: Update CategoryChips component to use onSelect instead of setSelectedCategory

The CategoryChips component has been updated to use the onSelect prop instead of the setSelectedCategory prop. This change improves the clarity and consistency of the component's API.

* feat: Add pagination functionality to ProductPicker component

The ProductPicker component now includes pagination functionality to limit the number of displayed products per page. This improvement allows for better performance and a more user-friendly experience when browsing and selecting products. The number of products per page is set to 8, and users can navigate between pages using the previous and next buttons. The total number of products and the current page number are also displayed for reference.

* feat: Add size attribute to Laporan button in FarmInputProductSales page

* fix: Fix error lint

* chore: Update npm dependencies to latest stable versions