Refactor getUserHashKey function and migrate user hash keys (#366)
This commit refactors the getUserHashKey function in the Auth provider. It introduces a new function, getUserHashKeyMainFn, which calculates the hash key using the sha3_256 algorithm. The old implementation, getUserHashKeyV1, is marked as deprecated.

Additionally, a new function, migrateUserHasKeyFromV1, is added to handle the migration of user hash keys from the old implementation to the new one. It retrieves the old hash key, retrieves the stored authentication information associated with it from the local storage, and then stores it with the new hash key. Finally, it removes the old hash key from the local storage.

This commit improves the code structure and prepares for future enhancements in the authentication functionality.

fix #364