Upgrades mui-datatable-delight dependency (#823)
* Upgrades mui-datatable-delight dependency

Updates the `mui-datatable-delight` package to `2.0.0-experimental.2`.

This upgrade brings various dependency updates, including newer versions of `@mui` and `@emotion` components, and aligns the `react` and `react-dom` peer dependencies to a specific version for improved compatibility.

* Updates `mui-datatable-delight` dependency

Upgrades the `mui-datatable-delight` package to `2.0.0-experimental.7`. This update incorporates the latest features and fixes from the library, and includes necessary adjustments to its internal dependencies, such as the addition of `redux`.