Moves online presence tracking component
Relocates the `PresenceOnlineUsers` component from `components/auth-layout` to `app/_parts`. This component is now rendered directly within the `RootLayout`.

Refactors the Ably client instantiation within the component to occur inside a `useEffect` hook, dependent on user authentication information. This ensures the client is initialized and managed correctly throughout its lifecycle, activating only when a user is authenticated and in a production environment.