Moves online presence tracking component (#853)
* Moves online presence tracking component

Relocates the `PresenceOnlineUsers` component from `components/auth-layout` to `app/_parts`. This component is now rendered directly within the `RootLayout`.

Refactors the Ably client instantiation within the component to occur inside a `useEffect` hook, dependent on user authentication information. This ensures the client is initialized and managed correctly throughout its lifecycle, activating only when a user is authenticated and in a production environment.

* Fix build error by disabling SSR for presence online users

Prevents an error where the presence channel could not be attached during server-side rendering. By dynamically importing the component with SSR disabled, it ensures the component only renders on the client.

fix #852
fix EKBS-NEXTJS-8J
https://sensasi-apps.sentry.io/issues/6984090342